Implement Picture portion of create post - Refactored ImangeUploadButton for reuse
