+++ application/src/components/components.scss
@@ -98,6 +98,17 @@ div.tag-container {
     }
 }
 
+.img-preview {
+    text-align: right;
+    
+    img {
+        max-width: 100%;
+        max-height: 250px;
+        object-fit: contain;
+        border-radius: 20px;
+    }
+}
+
 /* IDs */
 #avatar-preview {
     max-width: calc(3.875rem + 0.625rem);
